Add files via upload
The task of the product manager is to identify how much time it takes for a user to make a purchase on your website. Your PM would like to see the users' duration from first arriving on the website on any given day until their first purchase on that day. Your final result should show the duration dynamic daily.